fix: 🐛 Improve info box data fetcher for techniques
Resolves #13. The previous implementation matched text in the entire
row, so if a field had a title in the text (like Toxic Plume which has
'Targets' in the synergy effect data), it would get that before the
actual field we wanted. This has been changed so now we exclusively
match on the field names, and also we check for a complete match rather
than it just appearing as included text.